requirements.txt
anthropic==0.43.1
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==1.12.3
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==7.4.4
pytest-asyncio==0.23.3
